Add ar dmp for debugging. Ignore mismatched library signatures
Add FunctionFixpointBase

Fix detection of uninitialized dynamically allocated variables; add some debugging aids

Improve filtering algorithm and options:

1. analyzer.py:
- Add separate --no-filter-functions option to control enabling
  of function filtering, so that a reasonable default can be
  given when filtering.
- Use "(std|__gnu_cxx)::" as default filter
- Modify help show the defaults are shown
2. function_fixpoint_base.cpp
- Added remove_return_type function to remove leading return type
  from demangled names to avoid filtering user functions that
  might return a type that would be filtered out.
3. ikos_analyzer.cpp
- Remove default function-filter. Instead the python script should
  supply the filter.

Comment out LLVM dump call for when using release LLVM

Modify detection of uninitialized variable in response to review and further testing.

1. Use ZNumbers for bit sets rather than unordered_sets to avoid
bound limitations. There still is a limitation in that we only
detect uninitialed reads if the number of bytes being read
does not exceed the maximum unsigned 32 bit number. This is
a limitation of the ZNumber implementation.
Some new utility functions were added for this purpose,
single_mask, double_mask, and make_clipped_mask.

2. Previously I had been assuming that lack of a cell meant
the memory is definitely uninitialized, but that needs to
indicate TOP (i.e. either initialized or uninitialized),
so instead now, if the allocation is of known size,
we create an uninitialized cell that covers the allocation.
The trick to do this is to write the "undefined" value to the memory.
If we read or write in the middle of such a cell we split it up so
that we still track the uninitialized parts.

3. I added some qualifiers to the dumping of MemoryLocations to
disambiguate them from variables.

4. In analyzer/src/checker/buffer_overflow.cpp I fixed a
copy-and-paste error.

5. The new stricter strategy caused problems for bitfield
operations as the underlying memory starts out undefined
and bits are only set using read/mask/write operations.
To allow these I exempt logical "And" and "Or" from
checking for uninitialized input, but only for the case
where one operand is constant. This should cover most
of the cases bitfields are being used.

6. Some of the existing analyzer regression tests needed to
be modified for the stricter uninitialized variable checking
since a failure results in an error that then makes
following code unreachable.

7. I couldn't figure out how to run all the analyzer regression
tests at once, so I added a simple script (runalltests)
to do this. I can remove this is you point out another way.

8. I added 18 new uva tests.

9. I moved translate_internal_function as suggested by the reviewer.

10. Other small changes suggested by the reviewer.